pyproject.toml
[tool.poetry]
name = "poetry-template"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
pydantic = {extras = ["dotenv"], version = "^1.10.13"}
psycopg2-binary = "^2.9.10"
pandas = "^2.3.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
taskipy = "^1.13.0"
pyinstaller = "^6.10.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "poetry run python src/main.py"
lint = "poetry run flake8 src --show-source --statistics"
format = "poetry run black src"
build = "poetry run pyinstaller -y src/main.py"
build-one = "poetry run pyinstaller -y src/main.py --onefile"
test = "poetry run pytest"
